refactor: 다이얼로그 textAlign에 따른 분류
