+++ lib/app.dart
@@ -2,6 +2,7 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:me_mind/chat/view/s_chat.dart';
 import 'package:me_mind/common/theme/custom_theme.dart';
+import 'package:me_mind/settings/view/s_setting.dart';
 import 'package:me_mind/user/view/s_signin.dart';
 import 'package:me_mind/user/view/s_signup_welcome.dart';
 import 'common/theme/custom_theme_app.dart';
@@ -47,7 +48,7 @@ class AppState extends State<App> with WidgetsBindingObserver {
           locale: context.locale,
           title: 'Me mind',
           theme: CustomThemeHolder.of(context).theme.themeData,
-          home: const SignUpWelcome(),
+          home: const Settings(),
         );
       }),
     );
