+++ lib/settings/view/s_setting.dart
@@ -1,7 +1,7 @@
 // ignore_for_file: sort_child_properties_last
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/svg.dart';
-import 'package:me_mind/common/component/custom_dialog.dart';
+import 'package:me_mind/common/component/dialog/d_common_dialog.dart';
+import 'package:me_mind/common/component/dialog/main.dart';
 import 'package:me_mind/common/constant/app_colors.dart';
 import 'package:me_mind/common/constant/font_sizes.dart';
 import 'package:me_mind/common/layout/default_layout.dart';
@@ -241,16 +241,15 @@ class _SettingState extends State<Settings> {
                     style: FontSizes.getHeadline2Style()
                         .copyWith(color: theme.appColors.iconButton),
                   ),
-                  onTap: () => getCustomDialog(context,
-                      buttonText: "네",
-                      cancelButtonText: "아니오",
-                      contentTitleText: "알림 수신 동의",
-                      isButtonWidthHalf: false,
-                      contentDetailText:
-                          "해당 기기로 서비스 이용에 필요한\n안내 사항을 푸시 알림으로\n보내드리겠습니다.\n\n앱 푸시에 수신 동의하시겠습니까?",
-                      textAlign: TextAlign.start, OnSubmit: () {
-                    Navigator.pop(context);
-                  }, onCancelSubmit: () {}),
+                  onTap: () => getDialog().getCustomDialog(context,
+                      buttonText: "아니오",
+                      imageLink: "assets/image/logo/logo.png",
+                      cancelButtonText: "네",
+                      contentTitleText: "정말 로그아웃 하시겠어요?",
+                      contentDetailText: "내년 1월 중 오픈 예정입니다.",
+                      isButtonWidthHalf: true,
+                      onSubmit: () {},
+                      onCancelSubmit: () {}),
                 ),
               )
             ],
